Refactor server components from stream to server
