+++ tests/stream/duplex.rs
@@ -5,7 +5,7 @@ async fn braided_duplex() {
 
     let (client, incoming) = hyperdriver::stream::duplex::pair("test".parse().unwrap());
 
-    let server = hyperdriver::stream::server::Acceptor::from(incoming);
+    let server = hyperdriver::server::conn::Acceptor::from(incoming);
     tokio::spawn(async move {
         let mut incoming = server.fuse();
         while let Some(Ok(mut stream)) = incoming.next().await {
