+++ tests/stream/tls.rs
@@ -42,7 +42,7 @@ async fn braided_tls() {
     let addr = incoming.local_addr().unwrap();
 
     let server =
-        hyperdriver::stream::server::Acceptor::from(incoming).with_tls(Arc::new(tls_config()));
+        hyperdriver::server::conn::Acceptor::from(incoming).with_tls(Arc::new(tls_config()));
 
     tokio::spawn(async move {
         let mut incoming = server.fuse();
