+++ tests/stream/unix.rs
@@ -7,7 +7,7 @@ async fn braided_unix() {
 
     let incoming = tokio::net::UnixListener::bind(dir.path().join("braid.sock")).unwrap();
 
-    let server = hyperdriver::stream::server::Acceptor::from(incoming);
+    let server = hyperdriver::server::conn::Acceptor::from(incoming);
     tokio::spawn(async move {
         let mut incoming = server.fuse();
         while let Some(Ok(mut stream)) = incoming.next().await {
